Moved function `test_crops_merge_gt` to test_merge.py in order to have a single test file per app
